Validate the domain name in the cookie against the public suffix
This prevents cookies from being set for domains that are not valid.

It also fixes the error:
  CookieJar::InvalidCookieError: Domain is inappropriate based on
  request URI hostname

This error was because the domain for the request could be a subdomain
of the domain in the cookie. For example, if the request was for
my.salesforce.com and the cookie domain was .salesforce.com,
the cookie would be valid. However, if the request was for
custom.sandbox.my.salesforce.com, the cookie would be invalid.

By using the public suffix, we can validate that the domain in the
cookie is valid for the request URI.